+++ src/file-resolver/index.ts
@@ -1,14 +1,8 @@
-import { RuleApplications, FileFilter } from './../types';
+import { FileFilter } from './../types';
 import { listFiles } from './../utils';
-import filterFilesToProcess from './filter-files-to-process';
 
-export default (
-    root: string,
-    exclude: FileFilter | undefined,
-    ruleApplications: RuleApplications,
-): string[] => {
+export default (root: string, exclude: FileFilter | undefined): string[] => {
     const fileList = listFiles(root, exclude);
-    const filesToProcess = filterFilesToProcess(root, fileList, ruleApplications);
 
-    return filesToProcess;
+    return fileList;
 };