+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Master
 
 - Expose TypeScript definitions
+- Parse all the files in the project instead of the relevant only (the result is cached)
 
 ## 0.3.0 - +1 rule, -1 external API